+++ src/components/Header/SubHeader/index.js
@@ -1,0 +1,1 @@
+export { default as SubHeader } from './SubHeader';
